Fix the z score and p value computation for SamplingSigClust
SamplingSigClust is made up of sub SigClusts. Previously the z score
was computed by concatenating all of the simulated CIs from all the
sub sigclusts, and then comparing the sample CI to that. But more appropriate is
to put the simulated CIs from the sub sigclusts on the same playing field by
first subtracting the sample CI from that sub sigclust. Then the p value is
the fraction of these differences that are negative.