Improve eigenvalue computation using Scipy
Previous method of computing eigenvalues used numpy.linalg.eigvals,
which is fine for d <= n. But for d > n the iterative method results
in complex eigenvalues and even with negative real parts. This can be fixed
with numpy.linalg.eigvalsh, but for d > n it is best to use the scipy routine
we are using now which specifically focuses on finding the large eigenvalues
(we know that d-n of them will be 0).